Add GaemBoard component
